+++ spark/sql-30/src/test/scala/org/opensearch/spark/sql/SchemaUtilsTest.scala
@@ -33,7 +33,7 @@ import java.util.{Map => JMap}
 import org.apache.spark.sql.types.ArrayType
 import org.apache.spark.sql.types.DataTypes._
 import org.apache.spark.sql.types.StructType
-import org.codehaus.jackson.map.ObjectMapper
+import com.fasterxml.jackson.databind.ObjectMapper
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue